Implement some service calls and Models

+++ src/main/java/com/example/sportsbetting/database/repository/IPlayerRepository.java
@@ -3,5 +3,5 @@ package com.example.sportsbetting.database.repository;
 import com.example.sportsbetting.database.model.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IPlayerRepository extends CrudRepository<Player, Long> {
+public interface IPlayerRepository extends CrudRepository<Player, Integer> {
 }
